refactor(Coin): remove enums package
